feat: Update backend and frontend components
- Modified authentication-related files in the backend: auth.controller.ts, auth.module.ts, auth.service.ts, and jwt.strategy.ts.
- Updated main.ts and task-related files: tasks.controller.ts and tasks.service.ts.
- Adjusted frontend components including sign-in, sign-up, task-form, task-list, and multi-select.
- Updated ESLint configuration and package.json in the frontend.
- Added a new utility file: axiosClient.ts.
- Removed obsolete migration files and AppHomeView.tsx.
- Updated various pages and routing files in the frontend.

This commit includes significant updates to both backend and frontend components to improve functionality and maintainability.

+++ packages/backend/src/tasks/tasks.service.ts
@@ -59,8 +59,27 @@ export class TasksService {
     };
   }
 
-  async deleteTask(taskId: number): Promise<void> {
+  async deleteTask(taskId: number, userId: number): Promise<void> {
     const sql = this.dbService.sql;
+
+    // Görev sahiplerini kontrol et
+    const owners = await sql`
+      SELECT
+        user_id
+      FROM
+        task_owners
+      WHERE
+        task_id = ${taskId}
+    `;
+
+    const ownerIds = owners.map((owner) => owner.user_id);
+
+    // Eğer kullanıcı görev sahibi değilse hata fırlat
+    if (!ownerIds.includes(userId)) {
+      throw new Error('You do not have permission to delete this task.');
+    }
+
+    // Görevi sil
     await sql`
       DELETE FROM task
       WHERE
